pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tiny-cats-model"
version = "0.1.0"
description = "A tiny image classifier for cats using PyTorch transfer learning"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "d-oit" },
]
requires-python = ">=3.10"
keywords = ["deep-learning", "pytorch", "image-classification", "cats", "transfer-learning"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "Pillow>=10.0.0",
  "PyYAML>=6.0",
  "tqdm>=4.66.0",
  "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "flake8>=6.1.0",
  "black>=23.9.0",
  "isort>=5.12.0",
  "mypy>=1.5.0",
  "jupyter>=1.0.0",
  "ipykernel>=6.25.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
]

[project.scripts]
train-cats = "train:main"
evaluate-cats = "evaluate:main"
predict-cats = "predict:main"

[project.urls]
Homepage = "https://github.com/d-oit/tiny-cats-model"
Repository = "https://github.com/d-oit/tiny-cats-model"
Issues = "https://github.com/d-oit/tiny-cats-model/issues"

# ---------------------------------------------------------------------------
# Tool configurations
# ---------------------------------------------------------------------------

[tool.setuptools.packages.find]
exclude = ["tests*", "notebooks*", "data*", "checkpoints*", "runs*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | notebooks
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_glob = ["notebooks/*", ".venv/*"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
exclude = ["notebooks", "build", "dist", "node_modules"]

[[tool.mypy.overrides]]
module = "torch.*"
follow_imports = "skip"

[[tool.mypy.overrides]]
module = "torchvision.*"
follow_imports = "skip"

[[tool.mypy.overrides]]
module = "PIL.*"
follow_imports = "skip"

[tool.ruff]
line-length = 88
exclude = ["notebooks", "node_modules", ".git"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E203", "E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["."]
omit = [
  "tests/*",
  "notebooks/*",
  "setup.py",
  ".venv/*",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if __name__ == .__main__.:",
  "raise NotImplementedError",
]
